Working parallel sample sort using MPI
I need to make some edits to the way I handle string output to the screen.  Right now even with moderate input array size, you get a segmentation fault (overflow) because of the way I'm handling them. I already have a fix but it was in an older version that I was playing with and didn't move the changes over yet. Still need to comment my code.  I added in quicksort instead of bubble sort.  I'm not doing a recursive sample sort. There really isn't a need once you break the array up into the number of processes.  Thought about playing with having the separate processes sample the input array and help generate the samples to determine splitting.